requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
motor==3.3.2
orjson==3.9.12
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
pymongo==4.6.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.26.0
watchfiles==0.21.0
websockets==12.0
